+++ RTNeural/gru/gru.tpp
@@ -5,7 +5,7 @@ namespace RTNeural
 
 #if !defined(USE_EIGEN) && !defined(USE_XSIMD) && !defined(USE_ACCELERATE)
 template <typename T>
-GRULayer<T>::GRULayer(size_t in_size, size_t out_size)
+GRULayer<T>::GRULayer(int in_size, int out_size)
     : Layer<T>(in_size, out_size)
     , zWeights(in_size, out_size)
     , rWeights(in_size, out_size)
@@ -18,7 +18,7 @@ GRULayer<T>::GRULayer(size_t in_size, size_t out_size)
 }
 
 template <typename T>
-GRULayer<T>::GRULayer(std::initializer_list<size_t> sizes)
+GRULayer<T>::GRULayer(std::initializer_list<int> sizes)
     : GRULayer<T>(*sizes.begin(), *(sizes.begin() + 1))
 {
 }
@@ -45,7 +45,7 @@ GRULayer<T>::~GRULayer()
 }
 
 template <typename T>
-GRULayer<T>::WeightSet::WeightSet(size_t in_size, size_t out_size)
+GRULayer<T>::WeightSet::WeightSet(int in_size, int out_size)
     : out_size(out_size)
 {
     W = new T*[out_size];
@@ -53,7 +53,7 @@ GRULayer<T>::WeightSet::WeightSet(size_t in_size, size_t out_size)
     b[0] = new T[out_size];
     b[1] = new T[out_size];
 
-    for(size_t i = 0; i < out_size; ++i)
+    for(int i = 0; i < out_size; ++i)
     {
         W[i] = new T[in_size];
         U[i] = new T[out_size];
@@ -66,7 +66,7 @@ GRULayer<T>::WeightSet::~WeightSet()
     delete[] b[0];
     delete[] b[1];
 
-    for(size_t i = 0; i < out_size; ++i)
+    for(int i = 0; i < out_size; ++i)
     {
         delete[] W[i];
         delete[] U[i];
@@ -79,9 +79,9 @@ GRULayer<T>::WeightSet::~WeightSet()
 template <typename T>
 void GRULayer<T>::setWVals(const std::vector<std::vector<T>>& wVals)
 {
-    for(size_t i = 0; i < Layer<T>::in_size; ++i)
+    for(int i = 0; i < Layer<T>::in_size; ++i)
     {
-        for(size_t k = 0; k < Layer<T>::out_size; ++k)
+        for(int k = 0; k < Layer<T>::out_size; ++k)
         {
             zWeights.W[k][i] = wVals[i][k];
             rWeights.W[k][i] = wVals[i][k + Layer<T>::out_size];
@@ -93,9 +93,9 @@ void GRULayer<T>::setWVals(const std::vector<std::vector<T>>& wVals)
 template <typename T>
 void GRULayer<T>::setWVals(T** wVals)
 {
-    for(size_t i = 0; i < Layer<T>::in_size; ++i)
+    for(int i = 0; i < Layer<T>::in_size; ++i)
     {
-        for(size_t k = 0; k < Layer<T>::out_size; ++k)
+        for(int k = 0; k < Layer<T>::out_size; ++k)
         {
             zWeights.W[k][i] = wVals[i][k];
             rWeights.W[k][i] = wVals[i][k + Layer<T>::out_size];
@@ -107,9 +107,9 @@ void GRULayer<T>::setWVals(T** wVals)
 template <typename T>
 void GRULayer<T>::setUVals(const std::vector<std::vector<T>>& uVals)
 {
-    for(size_t i = 0; i < Layer<T>::out_size; ++i)
+    for(int i = 0; i < Layer<T>::out_size; ++i)
     {
-        for(size_t k = 0; k < Layer<T>::out_size; ++k)
+        for(int k = 0; k < Layer<T>::out_size; ++k)
         {
             zWeights.U[k][i] = uVals[i][k];
             rWeights.U[k][i] = uVals[i][k + Layer<T>::out_size];
@@ -121,9 +121,9 @@ void GRULayer<T>::setUVals(const std::vector<std::vector<T>>& uVals)
 template <typename T>
 void GRULayer<T>::setUVals(T** uVals)
 {
-    for(size_t i = 0; i < Layer<T>::out_size; ++i)
+    for(int i = 0; i < Layer<T>::out_size; ++i)
     {
-        for(size_t k = 0; k < Layer<T>::out_size; ++k)
+        for(int k = 0; k < Layer<T>::out_size; ++k)
         {
             zWeights.U[k][i] = uVals[i][k];
             rWeights.U[k][i] = uVals[i][k + Layer<T>::out_size];
@@ -135,9 +135,9 @@ void GRULayer<T>::setUVals(T** uVals)
 template <typename T>
 void GRULayer<T>::setBVals(const std::vector<std::vector<T>>& bVals)
 {
-    for(size_t i = 0; i < 2; ++i)
+    for(int i = 0; i < 2; ++i)
     {
-        for(size_t k = 0; k < Layer<T>::out_size; ++k)
+        for(int k = 0; k < Layer<T>::out_size; ++k)
         {
             zWeights.b[i][k] = bVals[i][k];
             rWeights.b[i][k] = bVals[i][k + Layer<T>::out_size];
@@ -149,9 +149,9 @@ void GRULayer<T>::setBVals(const std::vector<std::vector<T>>& bVals)
 template <typename T>
 void GRULayer<T>::setBVals(T** bVals)
 {
-    for(size_t i = 0; i < 2; ++i)
+    for(int i = 0; i < 2; ++i)
     {
-        for(size_t k = 0; k < Layer<T>::out_size; ++k)
+        for(int k = 0; k < Layer<T>::out_size; ++k)
         {
             zWeights.b[i][k] = bVals[i][k];
             rWeights.b[i][k] = bVals[i][k + Layer<T>::out_size];
@@ -161,7 +161,7 @@ void GRULayer<T>::setBVals(T** bVals)
 }
 
 template <typename T>
-T GRULayer<T>::getWVal(size_t i, size_t k) const noexcept
+T GRULayer<T>::getWVal(int i, int k) const noexcept
 {
     T** set = zWeights.W;
     if(k > 2 * Layer<T>::out_size)
@@ -179,7 +179,7 @@ T GRULayer<T>::getWVal(size_t i, size_t k) const noexcept
 }
 
 template <typename T>
-T GRULayer<T>::getUVal(size_t i, size_t k) const noexcept
+T GRULayer<T>::getUVal(int i, int k) const noexcept
 {
     T** set = zWeights.U;
     if(k > 2 * Layer<T>::out_size)
@@ -197,7 +197,7 @@ T GRULayer<T>::getUVal(size_t i, size_t k) const noexcept
 }
 
 template <typename T>
-T GRULayer<T>::getBVal(size_t i, size_t k) const noexcept
+T GRULayer<T>::getBVal(int i, int k) const noexcept
 {
     T** set = zWeights.b;
     if(k > 2 * Layer<T>::out_size)
@@ -215,10 +215,10 @@ T GRULayer<T>::getBVal(size_t i, size_t k) const noexcept
 }
 
 //====================================================
-template <typename T, size_t in_sizet, size_t out_sizet>
+template <typename T, int in_sizet, int out_sizet>
 GRULayerT<T, in_sizet, out_sizet>::GRULayerT()
 {
-    for(size_t i = 0; i < out_size; ++i)
+    for(int i = 0; i < out_size; ++i)
     {
         // single-input kernel weights
         Wz_1[i] = (T)0;
@@ -238,10 +238,10 @@ GRULayerT<T, in_sizet, out_sizet>::GRULayerT()
         ht[i] = (T)0;
     }
 
-    for(size_t i = 0; i < out_size; ++i)
+    for(int i = 0; i < out_size; ++i)
     {
         // recurrent weights
-        for(size_t k = 0; k < out_size; ++k)
+        for(int k = 0; k < out_size; ++k)
         {
             Uz[i][k] = (T)0;
             Ur[i][k] = (T)0;
@@ -249,7 +249,7 @@ GRULayerT<T, in_sizet, out_sizet>::GRULayerT()
         }
 
         // kernel weights
-        for(size_t k = 0; k < in_size; ++k)
+        for(int k = 0; k < in_size; ++k)
         {
             Wz[i][k] = (T)0;
             Wr[i][k] = (T)0;
@@ -260,21 +260,21 @@ GRULayerT<T, in_sizet, out_sizet>::GRULayerT()
     reset();
 }
 
-template <typename T, size_t in_sizet, size_t out_sizet>
+template <typename T, int in_sizet, int out_sizet>
 void GRULayerT<T, in_sizet, out_sizet>::reset()
 {
     // reset output state
-    for(size_t i = 0; i < out_size; ++i)
+    for(int i = 0; i < out_size; ++i)
         outs[i] = (T)0;
 }
 
 // kernel weights
-template <typename T, size_t in_sizet, size_t out_sizet>
+template <typename T, int in_sizet, int out_sizet>
 void GRULayerT<T, in_sizet, out_sizet>::setWVals(const std::vector<std::vector<T>>& wVals)
 {
-    for(size_t i = 0; i < in_size; ++i)
+    for(int i = 0; i < in_size; ++i)
     {
-        for(size_t j = 0; j < out_size; ++j)
+        for(int j = 0; j < out_size; ++j)
         {
             Wz[j][i] = wVals[i][j];
             Wr[j][i] = wVals[i][j + out_size];
@@ -282,7 +282,7 @@ void GRULayerT<T, in_sizet, out_sizet>::setWVals(const std::vector<std::vector<T
         }
     }
 
-    for(size_t j = 0; j < out_size; ++j)
+    for(int j = 0; j < out_size; ++j)
     {
         Wz_1[j] = wVals[0][j];
         Wr_1[j] = wVals[0][j + out_size];
@@ -291,12 +291,12 @@ void GRULayerT<T, in_sizet, out_sizet>::setWVals(const std::vector<std::vector<T
 }
 
 // recurrent weights
-template <typename T, size_t in_sizet, size_t out_sizet>
+template <typename T, int in_sizet, int out_sizet>
 void GRULayerT<T, in_sizet, out_sizet>::setUVals(const std::vector<std::vector<T>>& uVals)
 {
-    for(size_t i = 0; i < out_size; ++i)
+    for(int i = 0; i < out_size; ++i)
     {
-        for(size_t j = 0; j < out_size; ++j)
+        for(int j = 0; j < out_size; ++j)
         {
             Uz[j][i] = uVals[i][j];
             Ur[j][i] = uVals[i][j + out_size];
@@ -306,10 +306,10 @@ void GRULayerT<T, in_sizet, out_sizet>::setUVals(const std::vector<std::vector<T
 }
 
 // biases
-template <typename T, size_t in_sizet, size_t out_sizet>
+template <typename T, int in_sizet, int out_sizet>
 void GRULayerT<T, in_sizet, out_sizet>::setBVals(const std::vector<std::vector<T>>& bVals)
 {
-    for(size_t k = 0; k < out_size; ++k)
+    for(int k = 0; k < out_size; ++k)
     {
         bz[k] = bVals[0][k] + bVals[1][k];
         br[k] = bVals[0][k + out_size] + bVals[1][k + out_size];
